+++ encryption-service/makefile
@@ -19,8 +19,8 @@ help:  ## Display this help
 
 ##### Files #####
 binary = encryption-service
-protobufs = enc/enc_grpc.pb.go enc/enc.pb.go authn/authn_grpc.pb.go authn/authn.pb.go authz/access_object.pb.go
-protosource = enc/enc.proto authn/authn.proto authz/access_object.proto
+protobufs = service/enc/enc_grpc.pb.go service/enc/enc.pb.go service/authn/authn_grpc.pb.go service/authn/authn.pb.go authz/access_object.pb.go
+protosource = service/enc/enc.proto service/authn/authn.proto authz/access_object.proto
 credentials = data/client.root.crt data/object_storage_key data/ca.crt data/client.root.key data/encryptonize.crt data/object_storage_id data/object_storage.crt
 coverage = coverage-unit.html coverage-e2e.html coverage-all.html
 
@@ -33,8 +33,8 @@ build: $(protobufs)  ## Build the Encryption Service binary
 	CGO_ENABLED=0 go build -v -tags "$(tags)" .
 
 $(protobufs): $(protosource)
-	protoc --go-grpc_out=enc --go_out=enc enc/enc.proto
-	protoc --go-grpc_out=authn --go_out=authn authn/authn.proto
+	protoc --go-grpc_out=enc --go_out=enc service/enc/enc.proto
+	protoc --go-grpc_out=authn --go_out=authn service/authn/authn.proto
 	protoc --go_out=authz authz/access_object.proto
 
 .PHONY: docker-build
