+++ encryption-service/service/authn/authn.go
@@ -24,7 +24,7 @@ import (
 
 	"encryption-service/contextkeys"
 	"encryption-service/crypt"
-	"encryption-service/health"
+	"encryption-service/service/health"
 	"encryption-service/interfaces"
 	log "encryption-service/logger"
 )
@@ -62,7 +62,7 @@ func (au *AuthnService) CheckAccessToken(ctx context.Context) (context.Context, 
 		return nil, err
 	}
 
-	// Don't authenticate health checks
+	// Don't authenticate service/health checks
 	// IMPORTANT! This check MUST stay at the top of this function
 	if methodName == health.HealthEndpointCheck || methodName == health.HealthEndpointWatch {
 		return ctx, nil
