+++ encryption-service/service/authn/user_handlers.go
