+++ encryption-service/service/enc/authz_handlers.go
