+++ encryption-service/service/enc/enc.go
