+++ encryption-service/service/enc/enc.proto
