+++ encryption-service/service/enc/permissions_handler.go
