+++ encryption-service/service/enc/permissions_handler_test.go
