+++ encryption-service/service/enc/storage_handlers.go
