+++ encryption-service/service/health/health.go
