+++ encryption-service/tests/grpc_e2e_tests/grpc_e2e_init_test.go
@@ -18,7 +18,7 @@ import (
 	"os"
 	"testing"
 
-	"encryption-service/authn"
+	"encryption-service/service/authn"
 )
 
 var endpoint = "127.0.0.1:9000"
